requirements.txt
acoular==21.5
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-generator==1.10
attrs==21.4.0
audioread==2.1.9
backcall==0.2.0
bleach==4.1.0
celluloid==0.2.0
certifi==2021.5.30
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.12
colorama==0.4.5
cycler
Cython==0.29.21
d2l
dataclasses==0.8
DateTime==4.5
decorator==4.4.2
defusedxml==0.7.1
entrypoints==0.4
et-xmlfile==1.1.0
future==0.18.2
graphviz==0.8.4
h5py==2.10.0
idna==2.6
imageio==2.15.0
imageio-ffmpeg==0.4.7
importlib-metadata==4.8.3
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.17.2
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.0.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.1
kiwisolver
librosa==0.9.2
llvmlite==0.31.0
MarkupSafe==2.0.1
matplotlib
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
moviepy==1.0.3
mxnet==1.7.0.post2
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.5
networkx==2.5.1
notebook==6.4.10
numba==0.48.0
numexpr==2.8.1
numpy==1.19.5
olefile==0.46
opencv-python==4.1.2.30
openpyxl==3.1.2
packaging==21.3
pandas
pandocfilters==1.5.0
parso==0.7.1
pickleshare==0.7.5
Pillow==8.4.0
pooch==1.6.0
proglog==0.1.10
prometheus-client==0.14.1
prompt-toolkit==3.0.30
pycparser==2.21
Pygments==2.12.0
pyparsing==3.0.9
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
pyroomacoustics==0.3.1
pyrsistent==0.18.0
PySide2==5.15.2.1
PySoundFile==0.9.0.post1
python-dateutil
pytz==2022.1
PyWavelets==1.1.1
pywin32==304
pywinpty==1.1.6
pyzmq==23.2.0
qtconsole==5.2.2
QtPy==2.0.1
requests==2.27.1
resampy==0.3.1
scikit-image==0.17.2
scikit-learn==0.22.1
scipy
seaborn
Send2Trash==1.8.0
shiboken2==5.15.2.1
sip==6.4.0
six
soundfile==0.11.0
sox==1.4.1
tables==3.7.0
termcolor==1.1.0
terminado==0.13.0
testpath==0.6.0
tifffile==2020.9.3
timm==0.6.5
toml==0.10.2
torch==1.10.2+cu113
torchaudio==0.10.2+cu113
torchstat==0.0.7
torchvision==0.11.3+cu113
tornado
tqdm==4.55.1
traitlets==4.3.3
traits==6.3.2
typing_extensions==4.1.1
urllib3==1.22
wcwidth==0.2.5
webencodings==0.5.1
whl==0.0.3
widgetsnbextension==3.6.1
wincertstore==0.2
xlrd
xlwt==1.3.0
zipp==3.6.0
zope.interface==5.4.0
